https://github.com/OpenIdentityPlatform/OpenAM-.Net-Agent/issues/3

+++ ru.org.openam.sdk/session/Response.cs
@@ -54,7 +54,7 @@ namespace ru.org.openam.sdk.session
         public String cdomain;
         public long maxtime;
         public long maxidle;
-        public byte maxcaching;
+        public long maxcaching;
         public long timeidle;
         public long timeleft;
         public state state;
@@ -82,7 +82,7 @@ namespace ru.org.openam.sdk.session
                         else if (attr.LocalName.Equals("maxidle"))
                             maxidle = long.Parse(attr.Value);
                         else if (attr.LocalName.Equals("maxcaching"))
-                            maxcaching = byte.Parse(attr.Value);
+                            maxcaching = long.Parse(attr.Value);
                         else if (attr.LocalName.Equals("timeidle"))
                             timeidle = long.Parse(attr.Value);
                         else if (attr.LocalName.Equals("timeleft"))
